Move common function into utils scripts for reuse
The obtainCharset function is used in a few places and to reduce duplicated code, this function has been moved to the utils script for reuse

OHM-827
